+++ VideoApp/VideoApp/Stores/Auth/InternalAuthStore.swift
@@ -20,7 +20,7 @@ import GoogleSignIn
 class InternalAuthStore: NSObject, AuthStoreWriting {
     weak var delegate: AuthStoreWritingDelegate?
     var isSignedIn: Bool { firebaseAuth.currentUser != nil }
-    var passcode: String? { fatalError("Passcode not supported by Firebase auth.") }
+    var passcode: String? { nil } // Not used for internal auth
     var userDisplayName: String { firebaseAuth.currentUser?.displayName ?? firebaseAuth.currentUser?.email ?? "Unknown" }
     private let api: APIConfiguring
     private let appSettingsStore: AppSettingsStoreWriting
